fixes to compile

+++ skfmm/distance_marcher_dinit.h
@@ -6,7 +6,7 @@
 class distanceMarcherDInit : public distanceMarcher
 {
 public:
-  distanceMarcherDInit(double *phi,      double *dx, long *flag,
+  distanceMarcherDInit(double *phi,      double *dx, long long *flag,
                        double *distance, int ndim,   int *shape,
                        bool self_test,   int order,  double narrow,
                        int periodic, double *dinit) :
